requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-extra-fields==3.7.0
drf-yasg==1.21.8
environs==14.1.1
filetype==1.2.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
marshmallow==3.26.1
multidict==6.1.0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
yarl==1.18.3